added temp (ajax) controller

+++ src/main/java/spring/jquery/demo/server/entity/Persistent.java
@@ -7,13 +7,15 @@ public interface Persistent {
 
     /**
      * Get the primary id of the persistent object.
+     *
      * @return the id
      */
-    public Integer getId();
+    public Long getId();
 
     /**
      * Set the id for the persistent object.
+     *
      * @param id id to set
      */
-    public void setId(Integer id);
+    public void setId(Long id);
 }
